refactor: Modify "type" property to be an array in hotel model and form
The "type" property in the hotel model and form has been modified to be an array of strings instead of a single string. This change allows for multiple types to be selected for a hotel. The backend code in `hotel.ts` and `types.ts` has been updated accordingly, and the frontend code in `ManageHotelForm.tsx` and `TypeSection.tsx` has been modified to handle the new "types" property. This change resolves the bug related to the "type" property and improves the functionality of the hotel management feature.